# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API documentation example",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://test.naufalrafif.com/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://staging.example.com/api",
            "description": "Staging Server"
        },
        {
            "url": "https://api.example.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/data-master/regency": {
            "get": {
                "tags": [
                    "Data Master - Regency"
                ],
                "summary": "Get all regency",
                "operationId": "c583cf6414d4188e4d2ffa7833f06d24",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by regency name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit data per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page data",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "province_id",
                        "in": "query",
                        "description": "Province id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "meta": {
                                            "properties": {
                                                "total_data": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total_pages": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "regency-1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Jakarta Pusat"
                                                    },
                                                    "alt_name": {
                                                        "type": "string",
                                                        "example": "Jakarta Pusat"
                                                    },
                                                    "latitude": {
                                                        "type": "string",
                                                        "example": "-6.21462"
                                                    },
                                                    "longitude": {
                                                        "type": "string",
                                                        "example": "106.84513"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Data Master - Regency"
                ],
                "summary": "Create a new regency",
                "operationId": "27447085d6d89724f25be274590fa133",
                "requestBody": {
                    "description": "Regency data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "province_id": {
                                        "type": "string",
                                        "example": "province-1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Jakarta Pusat"
                                    },
                                    "alt_name": {
                                        "type": "string",
                                        "example": "Jakarta Pusat"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "-6.21462"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "106.84513"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "regency-1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jakarta Pusat"
                                                },
                                                "alt_name": {
                                                    "type": "string",
                                                    "example": "Jakarta Pusat"
                                                },
                                                "latitude": {
                                                    "type": "string",
                                                    "example": "-6.21462"
                                                },
                                                "longitude": {
                                                    "type": "string",
                                                    "example": "106.84513"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}